feat(nvim): add gitsigns
